Add security checks for file importation

+++ src/Service/TempfileService.php
@@ -4,9 +4,15 @@ namespace App\Service;
 
 use App\Entity\Tempfile;
 use App\Entity\User;
+use App\Util\Exception\BackendException;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
+use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Validator\Constraints\File;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class TempfileService
 {
@@ -14,13 +20,24 @@ class TempfileService
 
     public function __construct(
         private EntityManagerInterface $em,
-        ParameterBagInterface $parameterBag
+        private Filesystem $filesystem,
+        ParameterBagInterface $parameterBag,
+        private Security $security,
+        private ValidatorInterface $validator
     ) {
         $this->config = $parameterBag->get('tempfile');
     }
 
     public function preload(UploadedFile $file, User $user): Tempfile
     {
+        $violations = $this->validator->validate($file, $this->getValidationConstraints());
+        if (count($violations) > 0) {
+            throw new BackendException(
+                statusCode: Response::HTTP_BAD_REQUEST,
+                message: 'tempfile.invalid_file',
+            );
+        }
+
         $tempfile = new Tempfile();
         $tempfile->setFilename($this->generateName());
         $tempfile->setFiletype($file->guessExtension());
@@ -38,8 +55,18 @@ class TempfileService
         return sprintf('%s/%s', $this->config['basePath'], $tempfile->getFilename());
     }
 
-    public function consumed(Tempfile $tempfile): void
+    public function consumed(Tempfile $tempfile, string $destination): void
     {
+        if ($this->security->getUser() !== $tempfile->getUser()) {
+            throw new BackendException(
+                statusCode: Response::HTTP_FORBIDDEN,
+                message: 'tempfile.forbidden',
+            );
+        }
+
+        $tempfilePath = $this->buildPath($tempfile);
+        $this->filesystem->rename($tempfilePath, $destination);
+
         $this->em->remove($tempfile);
         $this->em->flush();
     }
@@ -47,5 +74,15 @@ class TempfileService
     private function generateName(): string
     {
         return uniqid('tmp_');
+    }
+
+    private function getValidationConstraints(): array
+    {
+        return [
+            new File(
+                maxSize: $this->config['maxSize'],
+                mimeTypes: $this->config['allowedMimeTypes']
+            ),
+        ];
     }
 }
